requirements.txt
# Production dependencies will be added incrementally by milestone.
fastapi
uvicorn
pydantic
pillow==12.1.1
numpy==2.4.3
mypy
